requirements.txt
discord.py
firebase-admin
python-dotenv
humanize
Flask
pytesseract
asyncpg
aiosqlite
colorthief
openai
azure-ai-inference
azure-core
jishaku